requirements.txt
aiodns==4.0.0
aiofiles==25.1.0
aiogram==3.24.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp_socks==0.11.0
aiosignal==1.4.0
alembic==1.18.4
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.37.1
anyio==4.12.1
astor==0.8.1
asttokens==3.0.1
async-generator==1.10
asyncpg==0.31.0
attrs==25.4.0
blessed==1.27.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cssselect2==0.9.0
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
editor==1.6.6
executing==2.2.1
fastapi==0.128.0
fastuuid==0.14.0
filelock==3.20.3
fonttools==4.61.1
fpdf2==2.8.5
frozenlist==1.8.0
fsspec==2026.1.0
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.6
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.2.0
html2image==2.0.7
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.3
idna==3.11
importlib_metadata==8.7.1
inquirer==3.4.1
ipykernel==6.31.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.12.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
keyboard==0.13.5
kiwisolver==1.4.9
litellm==1.81.1
magic-filter==1.0.12
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mido==1.3.3
MouseInfo==0.1.3
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
open-interpreter==0.4.3
openai==2.15.0
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.1
parso==0.8.5
pefile==2024.8.26
phonenumbers==9.0.25
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==4.25.8
psutil==5.9.8
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycares==5.0.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.1
PyMsgBox==2.0.1
pyparsing==3.3.1
pypdf==6.7.3
pyperclip==1.11.0
pyphen==0.17.2
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-rtmidi==1.5.8
python-socks==2.8.1
pytweening==1.2.0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
pyzmq==27.1.0
readchar==4.2.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
runs==1.2.2
scapy==2.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
selenium==4.40.0
Send2Trash==1.8.3
setuptools==80.10.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
SQLAlchemy==2.0.48
stack-data==0.6.3
starlette==0.37.2
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tkinterdnd2==0.4.3
tokenizers==0.22.2
tokentrim==0.1.13
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer==0.12.5
typer-slim==0.21.1
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.3.0
weasyprint==68.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wget==3.2
wsproto==1.3.2
xmod==1.8.1
yarl==1.22.0
yaspin==3.4.0
zipp==3.23.0
zopfli==0.4.1
